+++ src/java/eus/tartangalh/crud/services/AlmacenFacadeREST.java
@@ -23,7 +23,7 @@ import javax.ws.rs.core.MediaType;
  * Servicio REST para gestionar entidades de tipo Almacen. Proporciona endpoints
  * para crear, actualizar, eliminar y consultar almacenes.
  *
- * Autor: Javier Martín Uría
+ * Autor: Andoni
  */
 @Path("eus.tartangalh.crud.create.almacen")
 public class AlmacenFacadeREST {
@@ -83,7 +83,7 @@ public class AlmacenFacadeREST {
      */
     @DELETE
     @Path("{id}")
-    public void borrarAlmacen(@PathParam("id") Long id) {
+    public void borrarAlmacen(@PathParam("id") Integer id) {
         try {
             LOGGER.log(Level.INFO, "Eliminando almacén {0}", id);
             ejb.borrarAlmacen(ejb.encontrarAlmacen(id));
@@ -103,7 +103,7 @@ public class AlmacenFacadeREST {
     @GET
     @Path("{id}")
     @Produces({MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON})
-    public Almacen encontrar(@PathParam("id") Long id) {
+    public Almacen encontrar(@PathParam("id") Integer id) {
         try {
             LOGGER.log(Level.INFO, "Leyendo datos del almacén {0}", id);
             return ejb.encontrarAlmacen(id);
